requirements.txt
confluent-kafka==2.4.0
ujson==5.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.38.0
streamlit-autorefresh==1.0.1
pandas==2.2.2
